only run supported metrics
